Use doubles instead of floats in example 5 to reflect Vitis being able to detect and implement single precision accumulators now

+++ example_5/xilinx/Entry.cpp
@@ -1,6 +1,6 @@
 #include "Example5.h"
 
-void Example5(float const A[], float const B[], float C[]) {
+void Example5(double const A[], double const B[], double C[]) {
   #pragma HLS INTERFACE m_axi port=A bundle=gmem0 offset=slave
   #pragma HLS INTERFACE m_axi port=B bundle=gmem1 offset=slave
   #pragma HLS INTERFACE m_axi port=C bundle=gmem2 offset=slave
@@ -11,7 +11,7 @@ void Example5(float const A[], float const B[], float C[]) {
   MatrixMultiplication(A, B, C);
 }
 
-void Example5_Reordered(float const A[], float const B[], float C[]) {
+void Example5_Reordered(double const A[], double const B[], double C[]) {
   #pragma HLS INTERFACE m_axi port=A bundle=gmem0 offset=slave
   #pragma HLS INTERFACE m_axi port=B bundle=gmem1 offset=slave
   #pragma HLS INTERFACE m_axi port=C bundle=gmem2 offset=slave
