+++ example_5/xilinx/Example5.h
@@ -4,5 +4,5 @@ constexpr int N = 32;
 constexpr int K = N;
 constexpr int M = N;
 
-void MatrixMultiplication(const float A[], const float B[], float C[]);
-void Entry(float const A[], float const B[], float C[]);
+void MatrixMultiplication(const double A[], const double B[], double C[]);
+void Entry(double const A[], double const B[], double C[]);
